added community form and some style
